--- Controls/_search/NewController/SearchController.wml
@@ -1,3 +1,0 @@
-<ws:partial template="{{_options.content}}"
-            on:search="_searchHandler()"
-            on:searchReset="_searchResetHandler()"/>
