--- Controls/_search/NewController/SearchInputContainer.wml
@@ -1,3 +1,0 @@
-<ws:partial template="{{_options.content}}"
-            on:valueChanged="_valueChanged()"
-            on:searchClick="_searchClick()"/>
