+++ Controls/_searchNew/Controller.ts
@@ -1,14 +1,20 @@
+import {ISearchOptions} from 'Controls/interface';
+import {NewSourceController} from 'Controls/dataSource';
 import {QueryWhereExpression} from 'Types/source';
 import {RecordSet} from 'Types/collection';
-import {ISearchControllerOptions, ISearchController} from './interface';
 
-export default class Controller implements ISearchController {
+export interface IControllerClassOptions extends ISearchOptions {
+   sourceController: NewSourceController;
+   searchValue?: string;
+}
 
-   protected _options: ISearchControllerOptions = null;
+export default class Controller {
+
+   protected _options: IControllerClassOptions = null;
 
    protected _searchValue: string = '';
 
-   constructor(options: ISearchControllerOptions) {
+   constructor(options: IControllerClassOptions) {
       this._options = options;
    }
 
@@ -31,12 +37,12 @@ export default class Controller implements ISearchController {
    search(value: string): Promise<RecordSet> {
       const filter: QueryWhereExpression<unknown> = {...this._options.sourceController.getFilter()};
 
-      filter[this._options.searchParam] = this._searchValue = this._trim(value);
+      filter[this._options.searchParam] = this._searchValue = value;
 
       return this._updateFilter(filter);
    }
 
-   update(options: Partial<ISearchControllerOptions>): void {
+   update(options: Partial<IControllerClassOptions>): void {
       if (options.hasOwnProperty('searchValue')) {
          if (options?.searchValue !== this._searchValue) {
             if (options.searchValue) {
@@ -50,9 +56,5 @@ export default class Controller implements ISearchController {
          ...this._options,
          ...options
       };
-   }
-
-   private _trim(value: string): string {
-      return this._options.searchValueTrim && value ? value.trim() : value;
    }
 }
