+++ Controls/searchNew.ts
@@ -1,4 +1,4 @@
 export {default as Controller} from './_searchNew/Controller';
 export {default as SearchInputContainer} from './_searchNew/InputContainer';
 export {default as SearchContainer} from './_searchNew/Container';
-export {default as SearchDelay} from './_searchNew/SearchResolver';
+export {default as SearchDelay} from './_searchNew/SearchDelay';
